Adds linters that disallow walrus usage
This feature is very debatable. So, I guess it is a good idea to give a reader a choice.
When to use and when not to use this feature.

So, I have added a link to https://github.com/wemake-services/wemake-python-styleguide
that forbid to use walrus and several other new features.

There's also a simple plugin for flake8 https://github.com/asottile/flake8-walrus
for ones that just want this one thing.